+++ src/pages/EventDetails/components/EventRenderer.tsx
@@ -330,11 +330,6 @@ const EventRenderer = ({ data, preview = false }: EventRendererProps) => {
               Dette arrangementet håndhever ikke aktive prikker
             </Alert>
           )}
-          {!data.can_cause_strikes && (
-            <Alert severity='info' variant='outlined'>
-              Dette arrangementet gir ikke prikker om du melder deg av for sent eller ikke møter opp
-            </Alert>
-          )}
         </>
       )}
       <ApplyInfo />